+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @folionote/constants
 
+## 0.2.6
+
+### Patch Changes
+
+- feat: ✨ implement advanced search functionality with filters and history
+- Updated dependencies
+  - @folionote/locales@0.2.6
+
 ## 0.2.5
 
 ### Patch Changes